requirements.txt
smbus2
RPi.bme280
paho-mqtt
